+++ app/src/androidTest/java/com/easybill/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.easybill
+package com.easybill
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
