+++ app/src/androidTest/java/com/easybill/database/ItemTest.kt
@@ -1,4 +1,4 @@
-package database
+package com.easybill.database
 
 import android.content.Context
 import androidx.room.Room
