+++ app/src/main/java/com/easybill/ArchiveListAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.easybill
+package com.easybill
 
 import android.view.View
 import android.view.ViewGroup
@@ -6,7 +6,7 @@ import android.widget.ArrayAdapter
 import android.widget.TextView
 import androidx.fragment.app.FragmentActivity
 
-class ArchivListAdapter(
+class ArchiveListAdapter(
     private val context: FragmentActivity,
     private val company: Array<String>,
     private val date: Array<String>,
