+++ app/src/main/java/com/easybill/database/BillWithItems.kt
@@ -1,8 +1,11 @@
-package database
+package com.easybill.database
 
 import androidx.room.Embedded
 import androidx.room.Relation
 
+/**
+ * A BillWithItems embeds a Bill and includes the Items of the bill.
+ */
 data class BillWithItems(
     @Embedded val bill: Bill,
