+++ app/src/main/java/com/easybill/database/Converters.kt
@@ -1,4 +1,4 @@
-package database
+package com.easybill.database
 
 import androidx.room.TypeConverter
 import java.time.Instant
@@ -6,7 +6,15 @@ import java.time.LocalDateTime
 import java.time.ZoneId
 import java.util.TimeZone
 
+/**
+ * Various converters to translate non-primitive types to a fitting
+ * database-representation.
+ */
 class Converters {
+
+    /**
+     * Converts a timestamp (milliseconds) to a LocalDateTime of the default time-zone.
+     */
     @TypeConverter
     fun localDateTimeFromTimestamp(value: Long?): LocalDateTime? {
         return value?.let {
@@ -17,6 +25,9 @@ class Converters {
         }
     }
 
+    /**
+     * Converts a LocalDateTime to a timestamp (milliseconds) of the default time-zone.
+     */
     @TypeConverter
     fun localDateTimeToTimestamp(value: LocalDateTime?): Long? {
         return value?.let {
