+++ app/src/main/java/com/easybill/fragments/ScanFragment.kt
@@ -1,4 +1,4 @@
-package com.example.easybill
+package com.easybill.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,7 +7,8 @@ import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import androidx.navigation.findNavController
-import com.example.easybill.databinding.ScanBinding
+import com.easybill.R
+import com.easybill.databinding.ScanBinding
 
 /**
  * A simple [Fragment] subclass.
