--- app/src/main/java/database/Item.kt
@@ -1,20 +1,0 @@
-package database
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity(tableName = "item")
-class Item(
-
-    @PrimaryKey(autoGenerate = true)
-    var id: Long = 0L,
-
-    var billId: Long = 0L,
-
-    var name: String = "",
-
-    var amount: Double = 0.0,
-
-    var singlePrice: Double = 0.0
-
-)
