+++ app/src/main/res/navigation/navigation.xml
@@ -5,7 +5,7 @@
 
     <fragment
         android:id="@+id/archiveFragment"
-        android:name="com.example.easybill.ArchiveFragment"
+        android:name="com.easybill.fragments.ArchiveFragment"
         android:label="ArchiveFragment" >
         <action
             android:id="@+id/action_archiveFragment_to_scanFragment"
@@ -22,7 +22,7 @@
     </fragment>
     <fragment
         android:id="@+id/scanFragment"
-        android:name="com.example.easybill.ScanFragment"
+        android:name="com.easybill.fragments.ScanFragment"
         android:label="ScanFragment" >
         <action
             android:id="@+id/action_scanFragment_to_archiveFragment"
@@ -30,7 +30,7 @@
     </fragment>
     <fragment
         android:id="@+id/detailedBillFragment"
-        android:name="com.example.easybill.DetailedBillFragment"
+        android:name="com.easybill.fragments.DetailedBillFragment"
         android:label="DetailedBillFragment" >
         <action
             android:id="@+id/action_detailedBillFragment_to_archiveFragment"
@@ -41,7 +41,7 @@
     </fragment>
     <fragment
         android:id="@+id/filterFragment"
-        android:name="com.example.easybill.FilterFragment"
+        android:name="com.easybill.fragments.FilterFragment"
         android:label="FilterFragment" >
         <action
             android:id="@+id/action_filterFragment_to_archiveFragment"
@@ -49,7 +49,7 @@
     </fragment>
     <fragment
         android:id="@+id/statisticsFragment"
-        android:name="com.example.easybill.StatisticsFragment"
+        android:name="com.easybill.fragments.StatisticsFragment"
         android:label="StatisticsFragment" >
         <action
             android:id="@+id/action_statisticsFragment_to_archiveFragment"
