+++ local.properties
@@ -4,5 +4,5 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-#Wed Jun 03 09:46:34 CEST 2020
-sdk.dir=C\:\\Users\\Thomas.DESKTOP-OQNA4AK\\AppData\\Local\\Android\\Sdk
+#Thu Jun 04 12:49:25 CEST 2020
+sdk.dir=/home/octav/Android/Sdk
